# server/src/main/resources/openapi/storyblock-v1.yaml
openapi: 3.1.0
info:
  title: StoryBlock Engine API
  version: 1.1.0
  description: Immutable, evidence-bound narrative editing contracts.
servers:
  - url: /v1
security:
  - bearerAuth: []
...
paths:
...
  /novels:
    post:
      operationId: createNovel
      summary: Create a novel from an initial canonical revision
      x-required-scopes: ["novel:admin"]
      parameters:
        - $ref: "#/components/parameters/IfMatchForCreation"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNovelRequest"
      responses:
        "201":
          description: Novel created
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NovelHead"
        default:
          $ref: "#/components/responses/Problem"
...
  /novels/{novelId}/pdf-renders:
    post:
      operationId: renderRevisionPdf
      summary: Render a complete immutable revision as a deterministic PDF
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PdfRenderRequest"
      responses:
        "200":
          description: Deterministic A4 PDF containing narrative text and image blocks
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Content-Disposition:
              $ref: "#/components/headers/ContentDisposition"
            X-PDF-Renderer-Version:
              schema:
                const: pdf-renderer-1.0.0
            X-PDF-Page-Count:
              schema:
                type: integer
                minimum: 1
            X-PDF-Image-Count:
              schema:
                type: integer
                minimum: 0
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/Problem"
...
  /novels/{novelId}/commits:
    post:
      operationId: commitEdit
      summary: Commit one validated operation with atomic head CAS
      x-required-scopes: ["novel:commit"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitRequest"
      responses:
        "200":
          description: Existing idempotent commit result
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitResponse"
        "201":
          description: New immutable revision committed
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitResponse"
        default:
          $ref: "#/components/responses/Problem"
...
  /novels/{novelId}/monitor-packets:
    post:
      operationId: createMonitorPacket
      summary: Resolve a bounded read-only packet for one monitor target
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorPacketRequest"
      responses:
        "200":
          description: Bounded render, detector context, and local invariants
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorPacket"
        default:
          $ref: "#/components/responses/Problem"
  /novels/{novelId}/monitor-runs:
    post:
      operationId: submitMonitorOutput
      summary: Persist an evidence-bound finding or inert proposed operation
      x-required-scopes: ["monitor:submit"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorSubmissionRequest"
      responses:
        "200":
          description: Existing idempotent monitor submission
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorSubmissionResponse"
        "201":
          description: Immutable monitor output created
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorSubmissionResponse"
        default:
          $ref: "#/components/responses/Problem"
  /novels/{novelId}/monitor-runs/{monitorRunId}:
    get:
      operationId: getMonitorRun
      summary: Get an immutable monitor output with derived stale state
      x-required-scopes: ["novel:read"]
      parameters:
        - $ref: "#/components/parameters/NovelId"
        - $ref: "#/components/parameters/MonitorRunId"
      responses:
        "200":
          description: Monitor output and current invalidation state
          headers:
            ETag:
              $ref: "#/components/headers/StrongEtag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonitorRunStatus"
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: nv_<public-key-id>.<256-bit-secret>
  parameters:
    NovelId:
      name: novelId
      in: path
      required: true
      schema:
        type: string
        pattern: '^nov_[0-9a-f-]{36}$'
...
    MonitorRunId:
      name: monitorRunId
      in: path
      required: true
      schema:
        type: string
        pattern: '^mrun_[0-9a-f-]{36}$'
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: A single quoted strong canonical SHA-256 ETag.
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
    IfMatchForCreation:
      name: If-Match
      in: header
      required: true
      description: Wildcard precondition for creation in a collection.
      schema:
        const: '*'
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 200
  headers:
    StrongEtag:
      description: Quoted strong ETag of the canonical revision or immutable resource view.
      required: true
      schema:
        type: string
        pattern: '^"sha256:[0-9a-f]{64}"$'
    Location:
      description: Absolute or origin-relative durable resource URI.
      required: true
      schema:
        type: string
        format: uri-reference
...
    ContentDisposition:
      description: Attachment filename for offline use.
      required: true
      schema:
        type: string
  responses:
    Problem:
      description: RFC problem response with a stable machine code
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      required: [type, title, status, code, detail, instance, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,63}$'
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        current_revision_id:
          type: string
        current_etag:
          type: string
      additionalProperties: true
    CreateNovelRequest:
      type: object
      required: [initial_revision]
      properties:
        initial_revision:
          $ref: "#/components/schemas/CanonicalRevision"
      additionalProperties: false
...
    NovelHead:
      type: object
      required: [novel_id, head_revision_id, head_sequence, head_hash, schema_version]
      properties:
        novel_id:
          type: string
        head_revision_id:
          type: string
        head_sequence:
          type: integer
          minimum: 0
        head_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        schema_version:
          type: string
      additionalProperties: false
    CanonicalRevision:
      type: object
      required: [schema_version, novel_id, revision_id, parent_revision_id, chapters, created_at, content_hash]
      properties:
        schema_version:
          const: 1.0.0
        novel_id:
          type: string
        revision_id:
          type: string
        parent_revision_id:
          type: [string, "null"]
        chapters:
          type: array
          items:
            type: object
            additionalProperties: true
        created_at:
          type: string
          format: date-time
        content_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        extensions:
          type: object
          additionalProperties: true
      additionalProperties: false
...
    PdfRenderRequest:
      type: object
      required: [revision_id]
      properties:
        revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
      additionalProperties: false
...
    RenderPacket:
      type: object
      required: [novel_id, revision_id, revision_hash, renderer_version, range, rendered_text, blocks, resolved_meta, offset_map, scene_boundaries]
      properties:
        novel_id:
          type: string
        revision_id:
          type: string
        revision_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        renderer_version:
          type: string
        range:
          $ref: "#/components/schemas/RenderRange"
        rendered_text:
          type: string
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/RenderedBlock"
        resolved_meta:
          type: array
          items:
            $ref: "#/components/schemas/ResolvedBlockMetadata"
        offset_map:
          type: array
          items:
            $ref: "#/components/schemas/RenderOffset"
        scene_boundaries:
          type: array
          items:
            $ref: "#/components/schemas/ResolvedSceneBoundary"
      additionalProperties: false
    RenderRange:
      type: object
      required: [from_block_id, to_block_id]
      properties:
        from_block_id:
          type: [string, "null"]
        to_block_id:
          type: [string, "null"]
      additionalProperties: false
...
    EditPreviewRequest:
      type: object
      required: [operation, candidate_revision_id, candidate_created_at]
      properties:
        operation:
          $ref: "#/components/schemas/OperationEnvelope"
        candidate_revision_id:
          type: string
        candidate_created_at:
          type: string
          format: date-time
      additionalProperties: false
    CommitRequest:
      allOf:
        - $ref: "#/components/schemas/EditPreviewRequest"
...
    CommitResponse:
      type: object
      required: [revision_id, sequence, content_hash, operation_id, idempotent_replay]
      properties:
        revision_id:
          type: string
        sequence:
          type: integer
          minimum: 1
        content_hash:
          type: string
        operation_id:
          type: string
        idempotent_replay:
          type: boolean
      additionalProperties: false
...
    MonitorPacketRequest:
      type: object
      required: [revision_id, revision_hash, target_block_id, neighbor_count]
      properties:
        revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
        revision_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        target_block_id:
          type: string
          pattern: '^blk_[0-9a-f-]{36}$'
        neighbor_count:
          type: integer
          minimum: 1
          maximum: 2
      additionalProperties: false
    MonitorPacket:
      type: object
      required: [novel_id, revision_id, revision_hash, monitor_version, rule_version, detector_rule_version, target_block_id, neighbor_count, render_packet, detector_findings, local_invariants, allowed_tools]
      properties:
        novel_id:
          type: string
          pattern: '^nov_[0-9a-f-]{36}$'
        revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
        revision_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        monitor_version:
          type: string
        rule_version:
          type: string
        detector_rule_version:
          type: string
        target_block_id:
          type: string
          pattern: '^blk_[0-9a-f-]{36}$'
        neighbor_count:
          type: integer
          minimum: 1
          maximum: 2
        render_packet:
          $ref: "#/components/schemas/RenderPacket"
        detector_findings:
          type: array
          items:
            $ref: "#/components/schemas/DetectorFinding"
        local_invariants:
          $ref: "#/components/schemas/MonitorLocalInvariants"
        allowed_tools:
          type: array
          minItems: 2
          maxItems: 2
          prefixItems:
            - const: submit_finding
            - const: submit_proposed_operation
      additionalProperties: false
    MonitorLocalInvariants:
      type: object
      required: [affected_ids_must_be_within_window, evidence_required, max_block_graphemes, max_direct_speakers, max_sentences, min_sentences, target_block_version_id, target_scene_id, window_blocks]
      properties:
        affected_ids_must_be_within_window:
          const: true
        evidence_required:
          const: true
        max_block_graphemes:
          const: 100
        max_direct_speakers:
          const: 1
        max_sentences:
          const: 2
        min_sentences:
          const: 1
        target_block_version_id:
          type: string
          pattern: '^blv_[0-9a-f-]{36}$'
        target_scene_id:
          type: string
          pattern: '^scn_[0-9a-f-]{36}$'
        window_blocks:
          type: array
          minItems: 1
          maxItems: 5
          uniqueItems: true
          items:
            $ref: "#/components/schemas/MonitorBlockFingerprint"
      additionalProperties: false
...
    MonitorSubmissionRequest:
      type: object
      required: [revision_id, revision_hash, target_block_id, neighbor_count, rule_version, affected_block_ids, output]
      properties:
        revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
        revision_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        target_block_id:
          type: string
          pattern: '^blk_[0-9a-f-]{36}$'
        neighbor_count:
          type: integer
          minimum: 1
          maximum: 2
        rule_version:
          type: string
        affected_block_ids:
          type: array
          minItems: 1
          maxItems: 5
          uniqueItems: true
          contains:
            type: string
          items:
            type: string
            pattern: '^blk_[0-9a-f-]{36}$'
        output:
          oneOf:
            - $ref: "#/components/schemas/MonitorFindingOutput"
            - $ref: "#/components/schemas/MonitorProposedOperationOutput"
          discriminator:
            propertyName: kind
            mapping:
              finding: "#/components/schemas/MonitorFindingOutput"
              proposed_operation: "#/components/schemas/MonitorProposedOperationOutput"
      additionalProperties: false
...
    MonitorRunStatus:
      type: object
      required: [monitor_run_id, output_id, output_kind, novel_id, revision_id, revision_hash, target_block_id, neighbor_count, monitor_version, rule_version, affected_blocks, output, request_hash, submitted_at, state, stale_reasons, rebase_allowed]
      properties:
        monitor_run_id:
          type: string
          pattern: '^mrun_[0-9a-f-]{36}$'
        output_id:
          type: string
          pattern: '^(mis|mpr)_[0-9a-f-]{36}$'
        output_kind:
          type: string
          enum: [finding, proposed_operation]
        novel_id:
          type: string
          pattern: '^nov_[0-9a-f-]{36}$'
        revision_id:
          type: string
          pattern: '^rev_[0-9a-f-]{36}$'
        revision_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        target_block_id:
          type: string
          pattern: '^blk_[0-9a-f-]{36}$'
        neighbor_count:
          type: integer
          minimum: 1
          maximum: 2
        monitor_version:
          type: string
        rule_version:
          type: string
        affected_blocks:
          type: array
          minItems: 1
          maxItems: 5
          uniqueItems: true
          items:
            $ref: "#/components/schemas/MonitorBlockFingerprint"
        output:
          oneOf:
            - $ref: "#/components/schemas/MonitorFindingOutput"
            - $ref: "#/components/schemas/MonitorProposedOperationOutput"
        request_hash:
          type: string
          pattern: '^sha256:[0-9a-f]{64}$'
        submitted_at:
          type: string
          format: date-time
        state:
          type: string
          enum: [current, stale]
        stale_reasons:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [head_changed, affected_block_changed, affected_block_missing, rule_version_changed]
        rebase_allowed:
          const: false
        idempotent_replay:
          type: boolean
      additionalProperties: false
    MonitorSubmissionResponse:
      allOf:
        - $ref: "#/components/schemas/MonitorRunStatus"
        - type: object
          required: [idempotent_replay]
          properties:
            idempotent_replay:
              type: boolean
...